ref(tracing): Move types from web-vitals
