+++ packages/tracing/src/browser/web-vitals/types.ts
@@ -31,10 +31,6 @@ export interface Metric {
   // together and calculate a total.
   id: string;
 
-  // `false` if the value of the metric may change in the future,
-  // for the current page.
-  isFinal: boolean;
-
   // Any performance entries used in the metric value calculation.
   // Note, entries will be added to the array as the value changes.
   entries: PerformanceEntry[];
@@ -43,53 +39,3 @@ export interface Metric {
 export interface ReportHandler {
   (metric: Metric): void;
 }
-
-// http://wicg.github.io/netinfo/#navigatornetworkinformation-interface
-export interface NavigatorNetworkInformation {
-  readonly connection?: NetworkInformation;
-}
-
-// http://wicg.github.io/netinfo/#connection-types
-type ConnectionType = 'bluetooth' | 'cellular' | 'ethernet' | 'mixed' | 'none' | 'other' | 'unknown' | 'wifi' | 'wimax';
-
-// http://wicg.github.io/netinfo/#effectiveconnectiontype-enum
-type EffectiveConnectionType = '2g' | '3g' | '4g' | 'slow-2g';
-
-// http://wicg.github.io/netinfo/#dom-megabit
-type Megabit = number;
-// http://wicg.github.io/netinfo/#dom-millisecond
-type Millisecond = number;
-
-// http://wicg.github.io/netinfo/#networkinformation-interface
-interface NetworkInformation extends EventTarget {
-  // http://wicg.github.io/netinfo/#type-attribute
-  readonly type?: ConnectionType;
-  // http://wicg.github.io/netinfo/#effectivetype-attribute
-  readonly effectiveType?: EffectiveConnectionType;
-  // http://wicg.github.io/netinfo/#downlinkmax-attribute
-  readonly downlinkMax?: Megabit;
-  // http://wicg.github.io/netinfo/#downlink-attribute
-  readonly downlink?: Megabit;
-  // http://wicg.github.io/netinfo/#rtt-attribute
-  readonly rtt?: Millisecond;
-  // http://wicg.github.io/netinfo/#savedata-attribute
-  readonly saveData?: boolean;
-  // http://wicg.github.io/netinfo/#handling-changes-to-the-underlying-connection
-  onchange?: EventListener;
-}
-
-// https://w3c.github.io/device-memory/#sec-device-memory-js-api
-export interface NavigatorDeviceMemory {
-  readonly deviceMemory?: number;
-}
-
-export type NavigationTimingPolyfillEntry = Omit<
-  PerformanceNavigationTiming,
-  | 'initiatorType'
-  | 'nextHopProtocol'
-  | 'redirectCount'
-  | 'transferSize'
-  | 'encodedBodySize'
-  | 'decodedBodySize'
-  | 'toJSON'
->;
